api/requirements.txt
flask
flask-cors
requests
googlemaps
python-dotenv